+++ my-app/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/koya_sone/react/my-app/src/index.js":"1"},{"size":3475,"mtime":1607328143290,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ni9uca","/home/koya_sone/react/my-app/src/index.js",[]]
+[{"/home/koya_sone/react/my-app/src/index.js":"1","/home/koya_sone/react/my-app/src/components/game/game.js":"2","/home/koya_sone/react/my-app/src/components/board/board.js":"3","/home/koya_sone/react/my-app/src/components/function/calculateWinner.js":"4","/home/koya_sone/react/my-app/src/components/square/square.js":"5"},{"size":229,"mtime":1607409271970,"results":"6","hashOfConfig":"7"},{"size":2043,"mtime":1607409230770,"results":"8","hashOfConfig":"7"},{"size":962,"mtime":1607409219380,"results":"9","hashOfConfig":"7"},{"size":454,"mtime":1607407806511,"results":"10","hashOfConfig":"7"},{"size":165,"mtime":1607408019718,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ni9uca",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"/home/koya_sone/react/my-app/src/index.js",[],"/home/koya_sone/react/my-app/src/components/game/game.js",[],["23","24"],"/home/koya_sone/react/my-app/src/components/board/board.js",[],"/home/koya_sone/react/my-app/src/components/function/calculateWinner.js",[],"/home/koya_sone/react/my-app/src/components/square/square.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
